Extract project IO into useProjectIO hook
